Mejorando estructura biologica de carpetas
