+++ src/pages/memory_simulator/components/atoms/DeleteConfirmationModal.tsx
