+++ src/pages/memory_simulator/components/atoms/TitleComponent.tsx
@@ -1,15 +1,14 @@
 export function TitleComponent() {
     return (
-      <div className="text-center py-6">
-        <h1
-          className="
+        <div className="text-center py-6">
+            <h1
+                className="
             text-4xl sm:text-6xl font-extrabold tracking-wide uppercase text-neutral-900
             drop-shadow-[2px_2px_0_rgba(255,0,0,0.3)]
           "
-        >
-          Simulador <span className="text-red-500">Memoria</span>
-        </h1>
-      </div>
+            >
+                Simulador <span className="text-red-500">Memoria</span>
+            </h1>
+        </div>
     );
-  }
-  
+}
