+++ src/pages/memory_simulator/components/molecules/ObjectDetailsModal.tsx
