+++ src/routes/simulador/memoria.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { MemorySimulator } from "../../pages/simulator/MemorySimulator";
+import { MemorySimulator } from "../../pages/memory_simulator/MemorySimulator";
 
 export const Route = createFileRoute("/simulador/memoria")({
     component: RouteComponent,
