+++ src/shared/utils/draw/sequenceDrawActions.ts
